+++ src/Symfony/Component/Console/Tests/Input/ArgvInputTest.php
@@ -26,17 +26,17 @@ class ArgvInputTest extends TestCase
         $r = new \ReflectionObject($input);
         $p = $r->getProperty('tokens');
 
-        $this->assertEquals(['foo'], $p->getValue($input), '__construct() automatically get its input from the argv server variable');
+        $this->assertSame(['foo'], $p->getValue($input), '__construct() automatically get its input from the argv server variable');
     }
 
     public function testParseArguments()
     {
         $input = new ArgvInput(['cli.php', 'foo']);
         $input->bind(new InputDefinition([new InputArgument('name')]));
-        $this->assertEquals(['name' => 'foo'], $input->getArguments(), '->parse() parses required arguments');
+        $this->assertSame(['name' => 'foo'], $input->getArguments(), '->parse() parses required arguments');
 
         $input->bind(new InputDefinition([new InputArgument('name')]));
-        $this->assertEquals(['name' => 'foo'], $input->getArguments(), '->parse() is stateless');
+        $this->assertSame(['name' => 'foo'], $input->getArguments(), '->parse() is stateless');
     }
 
     /**
@@ -57,7 +57,7 @@ class ArgvInputTest extends TestCase
     {
         $input = new ArgvInput($input);
         $input->bind(new InputDefinition($options));
-        $this->assertEquals($expectedOptions, $input->getOptions(), $message);
+        $this->assertSame($expectedOptions, $input->getOptions(), $message);
     }
 
     public static function provideOptions()
@@ -363,7 +363,7 @@ class ArgvInputTest extends TestCase
         $input = new ArgvInput(['cli.php', 'foo', 'bar', 'baz', 'bat']);
         $input->bind(new InputDefinition([new InputArgument('name', InputArgument::IS_ARRAY)]));
 
-        $this->assertEquals(['name' => ['foo', 'bar', 'baz', 'bat']], $input->getArguments(), '->parse() parses array arguments');
+        $this->assertSame(['name' => ['foo', 'bar', 'baz', 'bat']], $input->getArguments(), '->parse() parses array arguments');
     }
 
     public function testParseArrayOption()
@@ -371,11 +371,11 @@ class ArgvInputTest extends TestCase
         $input = new ArgvInput(['cli.php', '--name=foo', '--name=bar', '--name=baz']);
         $input->bind(new InputDefinition([new InputOption('name', null, InputOption::VALUE_OPTIONAL | InputOption::VALUE_IS_ARRAY)]));
 
-        $this->assertEquals(['name' => ['foo', 'bar', 'baz']], $input->getOptions(), '->parse() parses array options ("--option=value" syntax)');
+        $this->assertSame(['name' => ['foo', 'bar', 'baz']], $input->getOptions(), '->parse() parses array options ("--option=value" syntax)');
 
         $input = new ArgvInput(['cli.php', '--name', 'foo', '--name', 'bar', '--name', 'baz']);
         $input->bind(new InputDefinition([new InputOption('name', null, InputOption::VALUE_OPTIONAL | InputOption::VALUE_IS_ARRAY)]));
-        $this->assertEquals(['name' => ['foo', 'bar', 'baz']], $input->getOptions(), '->parse() parses array options ("--option value" syntax)');
+        $this->assertSame(['name' => ['foo', 'bar', 'baz']], $input->getOptions(), '->parse() parses array options ("--option value" syntax)');
 
         $input = new ArgvInput(['cli.php', '--name=foo', '--name=bar', '--name=']);
         $input->bind(new InputDefinition([new InputOption('name', null, InputOption::VALUE_OPTIONAL | InputOption::VALUE_IS_ARRAY)]));
@@ -393,12 +393,12 @@ class ArgvInputTest extends TestCase
     {
         $input = new ArgvInput(['cli.php', '--', '-1']);
         $input->bind(new InputDefinition([new InputArgument('number')]));
-        $this->assertEquals(['number' => '-1'], $input->getArguments(), '->parse() parses arguments with leading dashes as arguments after having encountered a double-dash sequence');
+        $this->assertSame(['number' => '-1'], $input->getArguments(), '->parse() parses arguments with leading dashes as arguments after having encountered a double-dash sequence');
 
         $input = new ArgvInput(['cli.php', '-f', 'bar', '--', '-1']);
         $input->bind(new InputDefinition([new InputArgument('number'), new InputOption('foo', 'f', InputOption::VALUE_OPTIONAL)]));
-        $this->assertEquals(['foo' => 'bar'], $input->getOptions(), '->parse() parses arguments with leading dashes as options before having encountered a double-dash sequence');
-        $this->assertEquals(['number' => '-1'], $input->getArguments(), '->parse() parses arguments with leading dashes as arguments after having encountered a double-dash sequence');
+        $this->assertSame(['foo' => 'bar'], $input->getOptions(), '->parse() parses arguments with leading dashes as options before having encountered a double-dash sequence');
+        $this->assertSame(['number' => '-1'], $input->getArguments(), '->parse() parses arguments with leading dashes as arguments after having encountered a double-dash sequence');
     }
 
     public function testParseEmptyStringArgument()
@@ -406,7 +406,7 @@ class ArgvInputTest extends TestCase
         $input = new ArgvInput(['cli.php', '-f', 'bar', '']);
         $input->bind(new InputDefinition([new InputArgument('empty'), new InputOption('foo', 'f', InputOption::VALUE_OPTIONAL)]));
 
-        $this->assertEquals(['empty' => ''], $input->getArguments(), '->parse() parses empty string arguments');
+        $this->assertSame(['empty' => ''], $input->getArguments(), '->parse() parses empty string arguments');
     }
 
     public function testGetFirstArgument()
@@ -415,7 +415,7 @@ class ArgvInputTest extends TestCase
         $this->assertNull($input->getFirstArgument(), '->getFirstArgument() returns null when there is no arguments');
 
         $input = new ArgvInput(['cli.php', '-fbbar', 'foo']);
-        $this->assertEquals('foo', $input->getFirstArgument(), '->getFirstArgument() returns the first argument from the raw input');
+        $this->assertSame('foo', $input->getFirstArgument(), '->getFirstArgument() returns the first argument from the raw input');
 
         $input = new ArgvInput(['cli.php', '--foo', 'fooval', 'bar']);
         $input->bind(new InputDefinition([new InputOption('foo', 'f', InputOption::VALUE_OPTIONAL), new InputArgument('arg')]));
@@ -495,7 +495,7 @@ class ArgvInputTest extends TestCase
         // No warning thrown
         $this->assertFalse($input->hasParameterOption(['-m', '']));
 
-        $this->assertEquals('dev', $input->getParameterOption(['-e', '']));
+        $this->assertSame('dev', $input->getParameterOption(['-e', '']));
         // No warning thrown
         $this->assertFalse($input->getParameterOption(['-m', '']));
     }
@@ -503,10 +503,10 @@ class ArgvInputTest extends TestCase
     public function testToString()
     {
         $input = new ArgvInput(['cli.php', '-f', 'foo']);
-        $this->assertEquals('-f foo', (string) $input);
+        $this->assertSame('-f foo', (string) $input);
 
         $input = new ArgvInput(['cli.php', '-f', '--bar=foo', 'a b c d', "A\nB'C"]);
-        $this->assertEquals('-f --bar=foo '.escapeshellarg('a b c d').' '.escapeshellarg("A\nB'C"), (string) $input);
+        $this->assertSame('-f --bar=foo '.escapeshellarg('a b c d').' '.escapeshellarg("A\nB'C"), (string) $input);
     }
 
     /**
@@ -515,7 +515,7 @@ class ArgvInputTest extends TestCase
     public function testGetParameterOptionEqualSign($argv, $key, $default, $onlyParams, $expected)
     {
         $input = new ArgvInput($argv);
-        $this->assertEquals($expected, $input->getParameterOption($key, $default, $onlyParams), '->getParameterOption() returns the expected value');
+        $this->assertSame($expected, $input->getParameterOption($key, $default, $onlyParams), '->getParameterOption() returns the expected value');
     }
 
     public static function provideGetParameterOptionValues()
@@ -539,33 +539,33 @@ class ArgvInputTest extends TestCase
     {
         $input = new ArgvInput(['cli.php', '-']);
         $input->bind(new InputDefinition([new InputArgument('file')]));
-        $this->assertEquals(['file' => '-'], $input->getArguments(), '->parse() parses single dash as an argument');
+        $this->assertSame(['file' => '-'], $input->getArguments(), '->parse() parses single dash as an argument');
     }
 
     public function testParseOptionWithValueOptionalGivenEmptyAndRequiredArgument()
     {
         $input = new ArgvInput(['cli.php', '--foo=', 'bar']);
         $input->bind(new InputDefinition([new InputOption('foo', 'f', InputOption::VALUE_OPTIONAL), new InputArgument('name', InputArgument::REQUIRED)]));
-        $this->assertEquals(['foo' => null], $input->getOptions(), '->parse() parses optional options with empty value as null');
-        $this->assertEquals(['name' => 'bar'], $input->getArguments(), '->parse() parses required arguments');
+        $this->assertSame(['foo' => null], $input->getOptions(), '->parse() parses optional options with empty value as null');
+        $this->assertSame(['name' => 'bar'], $input->getArguments(), '->parse() parses required arguments');
 
         $input = new ArgvInput(['cli.php', '--foo=0', 'bar']);
         $input->bind(new InputDefinition([new InputOption('foo', 'f', InputOption::VALUE_OPTIONAL), new InputArgument('name', InputArgument::REQUIRED)]));
-        $this->assertEquals(['foo' => '0'], $input->getOptions(), '->parse() parses optional options with empty value as null');
-        $this->assertEquals(['name' => 'bar'], $input->getArguments(), '->parse() parses required arguments');
+        $this->assertSame(['foo' => '0'], $input->getOptions(), '->parse() parses optional options with empty value as null');
+        $this->assertSame(['name' => 'bar'], $input->getArguments(), '->parse() parses required arguments');
     }
 
     public function testParseOptionWithValueOptionalGivenEmptyAndOptionalArgument()
     {
         $input = new ArgvInput(['cli.php', '--foo=', 'bar']);
         $input->bind(new InputDefinition([new InputOption('foo', 'f', InputOption::VALUE_OPTIONAL), new InputArgument('name', InputArgument::OPTIONAL)]));
-        $this->assertEquals(['foo' => null], $input->getOptions(), '->parse() parses optional options with empty value as null');
-        $this->assertEquals(['name' => 'bar'], $input->getArguments(), '->parse() parses optional arguments');
+        $this->assertSame(['foo' => null], $input->getOptions(), '->parse() parses optional options with empty value as null');
+        $this->assertSame(['name' => 'bar'], $input->getArguments(), '->parse() parses optional arguments');
 
         $input = new ArgvInput(['cli.php', '--foo=0', 'bar']);
         $input->bind(new InputDefinition([new InputOption('foo', 'f', InputOption::VALUE_OPTIONAL), new InputArgument('name', InputArgument::OPTIONAL)]));
-        $this->assertEquals(['foo' => '0'], $input->getOptions(), '->parse() parses optional options with empty value as null');
-        $this->assertEquals(['name' => 'bar'], $input->getArguments(), '->parse() parses optional arguments');
+        $this->assertSame(['foo' => '0'], $input->getOptions(), '->parse() parses optional options with empty value as null');
+        $this->assertSame(['name' => 'bar'], $input->getArguments(), '->parse() parses optional arguments');
     }
 
     public function testGetRawTokensFalse()
@@ -593,5 +593,262 @@ class ArgvInputTest extends TestCase
         yield [['app/console', '--no-ansi', 'foo:bar', 'argument'], ['argument']];
         yield [['app/console', '--no-ansi', 'foo:bar', 'foo:bar'], ['foo:bar']];
         yield [['app/console', '--no-ansi', 'foo:bar', '--', 'argument'], ['--', 'argument']];
+    }
+
+    /**
+     * @dataProvider unparseProvider
+     */
+    public function testUnparse(
+        ?InputDefinition $inputDefinition,
+        ArgvInput $input,
+        ?array $parsedOptions,
+        array $expected,
+    ): void
+    {
+        if (null !== $inputDefinition) {
+            $input->bind($inputDefinition);
+        }
+
+        $actual = null === $parsedOptions ? $input->unparse() : $input->unparse($parsedOptions);
+
+        self::assertSame($expected, $actual);
+    }
+
+    public static function unparseProvider(): iterable
+    {
+        yield 'empty input and empty definition' => [
+            new InputDefinition(),
+            new ArgvInput([]),
+            [],
+            [],
+        ];
+
+        yield 'empty input and definition with default values: ignore default values' => [
+            new InputDefinition([
+                new InputArgument(
+                    'argWithDefaultValue',
+                    InputArgument::OPTIONAL,
+                    'Argument with a default value',
+                    'arg1DefaultValue',
+                ),
+                new InputOption(
+                    'optWithDefaultValue',
+                    null,
+                    InputOption::VALUE_REQUIRED,
+                    'Option with a default value',
+                    'opt1DefaultValue',
+                ),
+            ]),
+            new ArgvInput([]),
+            [],
+            [],
+        ];
+
+        $completeInputDefinition = new InputDefinition([
+            new InputArgument(
+                'requiredArgWithoutDefaultValue',
+                InputArgument::REQUIRED,
+                'Argument without a default value',
+            ),
+            new InputArgument(
+                'optionalArgWithDefaultValue',
+                InputArgument::OPTIONAL,
+                'Argument with a default value',
+                'argDefaultValue',
+            ),
+            new InputOption(
+                'optWithoutDefaultValue',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Option without a default value',
+            ),
+            new InputOption(
+                'optWithDefaultValue',
+                null,
+                InputOption::VALUE_REQUIRED,
+                'Option with a default value',
+                'optDefaultValue',
+            ),
+        ]);
+
+        yield 'arguments & options: returns all passed options but ignore default values' => [
+            $completeInputDefinition,
+            new ArgvInput(['argValue', '--optWithoutDefaultValue=optValue']),
+            [],
+            ['--optWithoutDefaultValue=optValue'],
+        ];
+
+        yield 'arguments & options; explicitly pass the default values: the default values are returned' => [
+            $completeInputDefinition,
+            new ArgvInput(['argValue', 'argDefaultValue', '--optWithoutDefaultValue=optValue', '--optWithDefaultValue=optDefaultValue']),
+            [],
+            [
+                '--optWithoutDefaultValue=optValue',
+                '--optWithDefaultValue=optDefaultValue',
+            ],
+        ];
+
+        yield 'arguments & options; no input definition: nothing returned' => [
+            null,
+            new ArgvInput(['argValue', 'argDefaultValue', '--optWithoutDefaultValue=optValue', '--optWithDefaultValue=optDefaultValue']),
+            [],
+            [],
+        ];
+
+        yield 'arguments & options; parsing an argument name instead of an option name: that option is ignored' => [
+            $completeInputDefinition,
+            new ArgvInput(['argValue']),
+            ['requiredArgWithoutDefaultValue'],
+            [],
+        ];
+
+        yield 'arguments & options; non passed option: it is ignored' => [
+            $completeInputDefinition,
+            new ArgvInput(['argValue']),
+            ['optWithDefaultValue'],
+            [],
+        ];
+
+        $createSingleOptionScenario = static fn (
+            InputOption $option,
+            array $input,
+            array $expected
+        ) => [
+            new InputDefinition([$option]),
+            new ArgvInput(['appName', ...$input]),
+            [],
+            $expected,
+        ];
+
+        yield 'option without value' => $createSingleOptionScenario(
+            new InputOption(
+                'opt',
+                null,
+                InputOption::VALUE_NONE,
+            ),
+            ['--opt'],
+            ['--opt'],
+        );
+
+        yield 'option without value by shortcut' => $createSingleOptionScenario(
+            new InputOption(
+                'opt',
+                'o',
+                InputOption::VALUE_NONE,
+            ),
+            ['-o'],
+            ['--opt'],
+        );
+
+        yield 'option with value required' => $createSingleOptionScenario(
+            new InputOption(
+                'opt',
+                null,
+                InputOption::VALUE_REQUIRED,
+            ),
+            ['--opt=foo'],
+            ['--opt=foo'],
+        );
+
+        yield 'option with non string value (bool)' => $createSingleOptionScenario(
+            new InputOption(
+                'opt',
+                null,
+                InputOption::VALUE_REQUIRED,
+            ),
+            ['--opt=1'],
+            ['--opt=1'],
+        );
+
+        yield 'option with non string value (int)' => $createSingleOptionScenario(
+            new InputOption(
+                'opt',
+                null,
+                InputOption::VALUE_REQUIRED,
+            ),
+            ['--opt=20'],
+            ['--opt=20'],
+        );
+
+        yield 'option with non string value (float)' => $createSingleOptionScenario(
+            new InputOption(
+                'opt',
+                null,
+                InputOption::VALUE_REQUIRED,
+            ),
+            ['--opt=5.3'],
+            ['--opt=\'5.3\''],
+        );
+
+        yield 'option with non string value (array of strings)' => $createSingleOptionScenario(
+            new InputOption(
+                'opt',
+                null,
+                InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY,
+            ),
+            ['--opt=v1', '--opt=v2', '--opt=v3'],
+            ['--opt=v1--opt=v2--opt=v3'],
+        );
+
+        yield 'negatable option (positive)' => $createSingleOptionScenario(
+            new InputOption(
+                'opt',
+                null,
+                InputOption::VALUE_NEGATABLE,
+            ),
+            ['--opt'],
+            ['--opt'],
+        );
+
+        yield 'negatable option (negative)' => $createSingleOptionScenario(
+            new InputOption(
+                'opt',
+                null,
+                InputOption::VALUE_NEGATABLE,
+            ),
+            ['--no-opt'],
+            ['--no-opt'],
+        );
+
+        $createEscapeOptionTokenScenario = static fn (
+            string $optionValue,
+            ?string $expected
+        ) => [
+            new InputDefinition([
+                new InputOption(
+                    'opt',
+                    null,
+                    InputOption::VALUE_REQUIRED,
+                ),
+            ]),
+            new ArgvInput(['appName', '--opt='.$optionValue]),
+            [],
+            ['--opt='.$expected],
+        ];
+
+        yield 'escape token; string token' => $createEscapeOptionTokenScenario(
+            'foo',
+            'foo',
+        );
+
+        yield 'escape token; escaped string token' => $createEscapeOptionTokenScenario(
+            '"foo"',
+            escapeshellarg('"foo"'),
+        );
+
+        yield 'escape token; escaped string token with both types of quotes' => $createEscapeOptionTokenScenario(
+            '"o_id in(\'20\')"',
+            escapeshellarg('"o_id in(\'20\')"'),
+        );
+
+        yield 'escape token; string token with spaces' => $createEscapeOptionTokenScenario(
+            'a b c d',
+            escapeshellarg('a b c d'),
+        );
+
+        yield 'escape token; string token with line return' => $createEscapeOptionTokenScenario(
+            "A\nB'C",
+            escapeshellarg("A\nB'C"),
+        );
     }
 }