+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Add `verbosity` argument to `mustRun` process helper method
  * [BC BREAK] Add silent verbosity (`--silent`/`SHELL_VERBOSITY=-2`) to suppress all output, including errors
  * Add `OutputInterface::isSilent()`, `Output::isSilent()`, `OutputStyle::isSilent()` methods
+ * Add `InputInterface::getRawArguments()`, `InputInterface::getRawOptions()` and `Input::unparse()` methods
 
 7.1
 ---